+++ CHANGELOG.md
@@ -1,9 +1,26 @@
 # Changelog
 
+## 1.0.4
+
+- [New Feature] Added GitLab integration
+
+## 1.0.3
+
+- [Bug Fix] Fixed `chariot test` command
+- [New Feature] Added Azure org-level integration
+- [New Feature] Now run scripts from any directory as plugins with --script flag
+
+## 1.0.2
+
+- [Bug Fix] Delete command was not making correct API calls.
+- [Bug Fix] Update command was unable to handle <KEY> argument and was throwing an error.
+- [Doc Update] For add command, clearer instructions are added for the `--key` argument.
+
 ## 1.0.1
+
 - Updated the `link` actions for a corresponding update in the backend API.
 
-
 ## 1.0.0
+
 Initial open-source release.